requirements.txt
ibm-db==3.1.4
prometheus-client==0.15.0
python-dotenv==0.21.0
pyyaml==6.0
asyncio==3.4.3
